apps/server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "The server app that provides RESTful API endpoints for this app"
authors = ["Ken Yip <ken20206@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8.17"
core = { path = "../../packages/core", develop = true }
flask-sqlalchemy = "^3.0.3"
flask = "^2.3.2"
pyjwt = "^2.7.0"
pydantic = "^1.10.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
